+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [CalVer](https://calver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+- Data: iyotetsu-bus の GTFS resource を 20260901 版へ更新。
+- Data: nishi-tokyo-bus の GTFS resource を 20260901 版へ更新。
+- Data: kawasaki-city-bus の GTFS resource を 20260828 版へ更新。
+- Data: meimon-taiyo-ferry の GTFS resource を 20260901 版へ更新。
+
 ## [2026.08.27]
 
 ### Changed